Add a drawLink function to remove duplicated code
