+++ src/imageTools/length.js
@@ -2,6 +2,7 @@ import external from '../externalModules.js';
 import mouseButtonTool from './mouseButtonTool.js';
 import touchTool from './touchTool.js';
 import drawTextBox from '../util/drawTextBox.js';
+import drawLink from '../util/drawLink.js';
 import toolStyle from '../stateManagement/toolStyle.js';
 import toolColors from '../stateManagement/toolColors.js';
 import drawHandles from '../manipulators/drawHandles.js';
@@ -188,51 +189,14 @@ function onImageRendered (e) {
 
     if (data.handles.textBox.hasMoved) {
       // Draw dashed link line between ellipse and text
-      const link = {
-        start: {},
-        end: {}
-      };
-
       const midpointCanvas = {
         x: (handleStartCanvas.x + handleEndCanvas.x) / 2,
         y: (handleStartCanvas.y + handleEndCanvas.y) / 2
       };
 
-      const points = [handleStartCanvas, handleEndCanvas, midpointCanvas];
+      const linkAnchorPoints = [handleStartCanvas, handleEndCanvas, midpointCanvas];
 
-      link.end.x = textCoords.x;
-      link.end.y = textCoords.y;
-
-      link.start = external.cornerstoneMath.point.findClosestPoint(points, link.end);
-
-      const boundingBoxPoints = [{
-        // Top middle point of bounding box
-        x: boundingBox.left + boundingBox.width / 2,
-        y: boundingBox.top
-      }, {
-        // Left middle point of bounding box
-        x: boundingBox.left,
-        y: boundingBox.top + boundingBox.height / 2
-      }, {
-        // Bottom middle point of bounding box
-        x: boundingBox.left + boundingBox.width / 2,
-        y: boundingBox.top + boundingBox.height
-      }, {
-        // Right middle point of bounding box
-        x: boundingBox.left + boundingBox.width,
-        y: boundingBox.top + boundingBox.height / 2
-      }
-      ];
-
-      link.end = external.cornerstoneMath.point.findClosestPoint(boundingBoxPoints, link.start);
-
-      context.beginPath();
-      context.strokeStyle = color;
-      context.lineWidth = lineWidth;
-      context.setLineDash([2, 3]);
-      context.moveTo(link.start.x, link.start.y);
-      context.lineTo(link.end.x, link.end.y);
-      context.stroke();
+      drawLink(linkAnchorPoints, textCoords, boundingBox, context, color, lineWidth);
     }
 
     context.restore();
