+++ src/imageTools/simpleAngle.js
@@ -3,6 +3,7 @@ import external from '../externalModules.js';
 import mouseButtonTool from './mouseButtonTool.js';
 import drawTextBox from '../util/drawTextBox.js';
 import roundToDecimal from '../util/roundToDecimal.js';
+import drawLink from '../util/drawLink.js';
 import textStyle from '../stateManagement/textStyle.js';
 import toolStyle from '../stateManagement/toolStyle.js';
 import toolColors from '../stateManagement/toolColors.js';
@@ -237,46 +238,9 @@ function onImageRendered (e) {
 
       if (data.handles.textBox.hasMoved) {
         // Draw dashed link line between tool and text
-        const link = {
-          start: {},
-          end: {}
-        };
+        const linkAnchorPoints = [handleStartCanvas, handleEndCanvas, handleMiddleCanvas];
 
-        const points = [handleStartCanvas, handleEndCanvas, handleMiddleCanvas];
-
-        link.end.x = textCoords.x;
-        link.end.y = textCoords.y;
-
-        link.start = external.cornerstoneMath.point.findClosestPoint(points, link.end);
-
-        const boundingBoxPoints = [{
-          // Top middle point of bounding box
-          x: boundingBox.left + boundingBox.width / 2,
-          y: boundingBox.top
-        }, {
-          // Left middle point of bounding box
-          x: boundingBox.left,
-          y: boundingBox.top + boundingBox.height / 2
-        }, {
-          // Bottom middle point of bounding box
-          x: boundingBox.left + boundingBox.width / 2,
-          y: boundingBox.top + boundingBox.height
-        }, {
-          // Right middle point of bounding box
-          x: boundingBox.left + boundingBox.width,
-          y: boundingBox.top + boundingBox.height / 2
-        }
-        ];
-
-        link.end = external.cornerstoneMath.point.findClosestPoint(boundingBoxPoints, link.start);
-
-        context.beginPath();
-        context.strokeStyle = color;
-        context.lineWidth = lineWidth;
-        context.setLineDash([2, 3]);
-        context.moveTo(link.start.x, link.start.y);
-        context.lineTo(link.end.x, link.end.y);
-        context.stroke();
+        drawLink(linkAnchorPoints, textCoords, boundingBox, context, color, lineWidth);
       }
     }
